+++ src_new/triggers/levi_ataxia/for_levi/leviticus/magi_offense_tracking/015_Erode.lua
@@ -22,7 +22,7 @@ attributes:
   isColorTriggerFg: 'no'
   isColorTriggerBg: 'no'
 triggerType: 0
-conditonLineDelta: 1
+conditonLineDelta: 3
 mStayOpen: 0
 mCommand: ''
 packageName: ''